requirements.txt
Flask
pandas
python-dotenv
groq
requests
